requirements.txt
sqlalchemy
uvicorn
fastapi
psycopg2-binary
python-dotenv